feat: create servico CRUD

+++ app/models/servico_model.py
@@ -1,6 +1,8 @@
 from core.configs import settings
 
-from sqlalchemy import Column, Integer, String, ForeignKey, Float, DECIMAL
+from sqlalchemy import Column, Integer, String, DECIMAL
+
+from decimal import Decimal
 
 
 class ServicoModel(settings.DBBaseModel):
@@ -8,5 +10,5 @@ class ServicoModel(settings.DBBaseModel):
  
     id: int = Column(Integer, primary_key=True, autoincrement=True)
     nome_servico: str = Column(String(100), nullable=False)
-    preco: float = Column(DECIMAL, nullable=False)
+    preco: Decimal = Column(DECIMAL, nullable=False)
     duracao_minutos: int = Column(Integer, nullable=False)
