+++ app/schemas/servico_schema.py
@@ -2,11 +2,13 @@ from typing import Optional
 
 from pydantic import BaseModel, ConfigDict
 
+from decimal import Decimal
+
 
 class ServicoSchema(BaseModel):
     id: Optional[int] = None
     nome_servico: str
-    preco: float
+    preco: Decimal
     duracao_minutos: int
 
     model_config = ConfigDict(from_attributes=True)
